feat!(cli): unify bootstrap/build-index into bootstrap and remove rag-build-index
