+++ src/local_rag_backend/cli_commands/__init__.py
@@ -6,12 +6,11 @@ from local_rag_backend.cli_commands.docs import (
     mutate_docs_cmd,
 )
 from local_rag_backend.cli_commands.eval import eval_cmd
-from local_rag_backend.cli_commands.index import build_index_cmd, rebuild_index_cmd, status_cmd
+from local_rag_backend.cli_commands.index import rebuild_index_cmd, status_cmd
 from local_rag_backend.cli_commands.server import server_cmd
 
 __all__ = [
     "bootstrap_cmd",
-    "build_index_cmd",
     "eval_cmd",
     "ingest_cmd",
     "mutate_docs_cmd",
